Works, with bugs

+++ Main/Snake.hpp
@@ -9,14 +9,14 @@ private:
 	sf::Vector2f tails_last_pos;
 	std::vector<sf::RectangleShape> segments; //probably rectangleshapes of a certain color
 
-	void add_segment(const int x_pos, const int y_pos, const sf::Color color);
+	void add_segment(const float x_pos, const float y_pos, const sf::Color color);
 	void add_segment(const sf::Vector2f& pos, const sf::Color color);
 public:
 	const enum class MOVEMENT_DIRECTIONS {
 		N, E, S, W
 	};
 
-	Snake(const int x_pos, const int y_pos);
+	Snake(const float x_pos, const float y_pos);
 	const std::vector<sf::RectangleShape>& get_segments() const {
 		return segments;
 	}
